# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos/upload-init": {
      "post": {
        "description": "Creates a draft video row in aguardando_upload state and starts a MinIO multipart upload. Returns the uploadId, bucket, key, and part size to use.",
        "operationId": "VideosController_initUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Multipart upload initialized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "videoId": {
                      "type": "string",
                      "example": "V1StGXR8_Z5jdHi6B-myT"
                    },
                    "uploadId": {
                      "type": "string"
                    },
                    "bucket": {
                      "type": "string",
                      "example": "streamtube-videos"
                    },
                    "key": {
                      "type": "string",
                      "example": "videos/{channelId}/{videoId}/source.mp4"
                    },
                    "partSize": {
                      "type": "number",
                      "example": 5242880
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Initialize a multipart upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{videoId}/upload-part-url": {
      "post": {
        "operationId": "VideosController_getPartUrl",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "nanoid-21 video identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadPartUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URL issued",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Caller does not own the video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not in aguardando_upload state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a presigned PUT URL for one part of the multipart upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{videoId}/upload-complete": {
      "post": {
        "operationId": "VideosController_complete",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "nanoid-21 video identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Multipart completed, job enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "videoId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "processando"
                    },
                    "queuedJobId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Caller does not own the video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not in aguardando_upload state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "S3 CompleteMultipartUpload rejected the part list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Finalize the multipart upload and enqueue the process-video job",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{videoId}/upload-abort": {
      "post": {
        "operationId": "VideosController_abort",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "nanoid-21 video identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Upload aborted"
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Caller does not own the video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not in aguardando_upload state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Abort the multipart upload and mark the video as erro",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{videoId}/stream": {
      "get": {
        "description": "Proxies a byte range from the storage backend. Honors the Range header; returns 206 with Content-Range/Accept-Ranges. Anonymous in Phase 03.",
        "operationId": "VideosController_stream",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "nanoid-21 video identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full content (no range)"
          },
          "206": {
            "description": "Partial content (range)"
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready for streaming",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "416": {
            "description": "Range header malformed or out of bounds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Stream a video (Range/206)",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{videoId}/download": {
      "get": {
        "description": "Returns the full source with Content-Disposition: attachment.",
        "operationId": "VideosController_download",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "nanoid-21 video identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full object streamed"
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready for download",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Download the full source object",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UploadInitDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "video/mp4",
              "video/quicktime",
              "video/webm"
            ],
            "example": "video/mp4"
          }
        },
        "required": [
          "mimeType"
        ]
      },
      "UploadPartUrlDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "minimum": 1,
            "maximum": 10000,
            "example": 1
          }
        },
        "required": [
          "partNumber"
        ]
      },
      "UploadCompletePartDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "minimum": 1,
            "maximum": 10000,
            "example": 1
          },
          "etag": {
            "type": "string",
            "example": "\"d41d8cd98f00b204e9800998ecf8427e\""
          }
        },
        "required": [
          "partNumber",
          "etag"
        ]
      },
      "UploadCompleteDto": {
        "type": "object",
        "properties": {
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadCompletePartDto"
            }
          }
        },
        "required": [
          "parts"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}